+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased] - ReleaseDate
 
+- Fixed parsing of HTTP headers with the same name. An HTTP client can
+  now send "Accept: image/gif\r\nAccept: image/jpeg\r\n" and Vial will
+  will return "image/gif, image/jpeg" from `request.header("Accept")`.
+
+  For more information see RFC2616:
+
+  https://greenbytes.de/tech/webdav/rfc2616.html#message.headers
+
 ## 0.1.7 (2020-10-21)
 
 - Added optional `cookies` feature.